+++ Changelog.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Beta User Acceptance Testing]
 </br>
 
+## [Unreleased] (2020-06-22)
+[test in qe](https://frontend-prs.cxengagelabs.net/qe/builds/skylight/1.48.11-SNAPSHOT/index.html)  
+[test in dev](https://frontend-prs.cxengagelabs.net/dev/builds/skylight/1.48.11-SNAPSHOT/index.html)
+### Changed
+* CXV1-22686 - Only do version check for refresh banner on load
+
 ## [1.48.10] (2020-06-18)
 [test in qe](https://frontend-prs.cxengagelabs.net/qe/builds/skylight/1.48.10/index.html)  
 [test in dev](https://frontend-prs.cxengagelabs.net/dev/builds/skylight/1.48.10/index.html)